requirements.txt
pandas
scikit_learn==1.6.0
botcity-maestro-sdk>=0.3.3,<1.0
